requirements.txt
certifi==2023.11.17
charset-normalizer==3.3.2
contourpy==1.2.0
cycler==0.12.1
distlib==0.3.8
filelock==3.13.1
fonttools==4.47.2
idna==3.6
kiwisolver==1.4.5
matplotlib==3.8.2
numpy==1.26.3
packaging==23.2
pillow==10.2.0
platformdirs==4.1.0
pyparsing==3.1.1
python-dateutil==2.8.2
requests==2.31.0
six==1.16.0
urllib3==2.1.0
virtualenv==20.25.0
